backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
httpx==0.26.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pyodbc>=5.0.1
alembic==1.13.1
mutagen==1.47.0
boto3==1.34.34
openai==1.12.0

